% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cloud_classification.R
\name{Cloud_Classification}
\alias{Cloud_Classification}
\title{Separate clouds into shallow cumulus and cumulus congestus}
\usage{
Cloud_Classification(data, CTHmax_s, CTHmin_c, CTHmax_c)
}
\arguments{
\item{data}{A dataframe containing all of the clouds and their properties}

\item{CTHmax_s}{The upper limit of cloud top height of shallow clouds}

\item{CTHmin_c}{The lower limit of cloud top height of congestus clouds}

\item{CTHmax_c}{The upper limit of cloud top height of congestus clouds}
}
\value{
two data frames (shallow and congestus), each containing the cloud properties (ER, CTH, CBH, thickness, RH) of
shallow and congestus clouds
}
\description{
Separate clouds into shallow cumulus and cumulus congestus
}
\examples{
data(clouds)
list(shallow, congestus) <- Cloud_Classification(clouds$CTH, 4000, 5000, 9000)
shallow
congestus

}
