% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouds.R
\docType{data}
\name{clouds}
\alias{clouds}
\title{Cloud Properties}
\format{An object of class \code{data.frame} with 787 rows and 5 columns.}
\usage{
data(clouds)
}
\description{
A dataset containing the entrainment rate (ER), relative humidity (RH), cloud top height (CTH),
cloud base height (CBH), and cloud depth (depth) of all of the clouds observed by a W-band 
cloud radar (WACR) during the GoAmazon 2014/5 campaign in Manacapuru, Brazil.
}
\keyword{datasets}
