% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_func.R
\name{Plot_func}
\alias{Plot_func}
\title{Make a scatter plot with trendlines of multiple datasets}
\usage{
Plot_func(data, varx, vary, cols, label, xlab, ylab, xlim, ylim, main)
}
\arguments{
\item{data}{A list of dataframes containing the variables being plotted}

\item{varx}{The column number of the x variable being plotted}

\item{vary}{The column number of the y variable being plotted}

\item{cols}{A vector of length of \code{data} containing the color of each dataset's points}

\item{label}{A vector of length of \code{data} list containing the label of each dataset}

\item{xlab}{The x axis label}

\item{ylab}{The y axis label}

\item{xlim}{The limits of the x axis}

\item{ylim}{The limits of the y axis}

\item{main}{The title of the plot}
}
\value{
Plot
}
\description{
Make a scatter plot with trendlines of multiple datasets
}
\examples{
data <- list(airquality[1:10,], airquality[11:20,], airquality[21:30,])
PlotER(data=data, varx=1, vary=3, cols=1:3, label=c("rows 1-10","rows 11-20","rows 21-30"),
       xlab="", ylab="", xlim="", ylim="", main = "")

}
