% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precip_classification.R
\name{Precip_Classification}
\alias{Precip_Classification}
\title{Separate clouds into precipitating and non-precipitating}
\usage{
Precip_Classification(cloud, CBH_thresh)
}
\arguments{
\item{cloud}{A dataframe containing the cloud properties}

\item{CBH_thresh}{The upper limit of cloud base height of precipitating clouds}
}
\value{
two data frames (precip and noprecip), each containing the cloud properties (ER, CTH, CBH, thickness, RH) of
precipitating and non-precipitating clouds
}
\description{
Separate clouds into precipitating and non-precipitating
}
\examples{
data(clouds)
precip <- Precip_Classification(clouds$CBH, 1000)$precip


}
