% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reiss.R
\docType{package}
\name{Reiss}
\alias{Reiss}
\alias{Reiss-package}
\title{Reiss: A package to organize functions for my EVRN 720 final project from the spring of 2018.}
\description{
Reiss: A package to organize functions for my EVRN 720 final project from the spring of 2018.
}
\section{Reiss functions}{

Cloud_Classification, Precip_Classification, Analyze, T_test, Hello, add
}

