% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_test.R
\name{T_test}
\alias{T_test}
\title{Find the difference in means of two variables}
\usage{
T_test(x, y)
}
\arguments{
\item{x}{A set of real numbers}

\item{y}{A set of real numbers}
}
\value{
the p- value of \code{x} and \code{y}
}
\description{
Find the difference in means of two variables
}
\examples{
T_test(airquality$Ozone, airquality$Wind)

}
